+++ core/src/test/java/com/ensody/reactivestate/SomeData.kt
@@ -1,0 +1,3 @@
+package com.ensody.reactivestate
+
+internal data class SomeData(val name: String)